% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{BestAccuracy}
\alias{BestAccuracy}
\title{\code{BestAccuracy} Asks for a dataframe, a vector of collumn indices and the goal collumn and returns the maximum possible value of accuracy of filling missing values}
\usage{
BestAccuracy(df, VECTORS, goal)
}
\arguments{
\item{df}{A dataframe that you intend to fill missing values, warning this dataframe shall contain no missing values so the user must drop the lines it happens}

\item{VECTORS}{The collumns you wish to use to predict the missing values}

\item{goal}{The collum with the missing values you wish to fill}
}
\description{
\code{BestAccuracy} Asks for a dataframe, a vector of collumn indices and the goal collumn and returns the maximum possible value of accuracy of filling missing values
}
