% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{BestVectorTable}
\alias{BestVectorTable}
\title{\code{BestVectorTable} Asks for a data.table and some parameters and returns the best combination of collums to predict the missing value}
\usage{
BestVectorTable(df, goal, maxi, repetitions, trigger = 1, ratio = 0.99)
}
\arguments{
\item{df}{A data.table with the missing values you wish to fill}

\item{goal}{The collum with the missing values you wish to fill}

\item{maxi}{What will be the length of possible combinations you will test example if 2 they will test up to all possible pairs of collums}

\item{repetitions}{Measure of error, the bigger the less likely you will get the right prediction}

\item{trigger}{When you pair all possible combination of tuples a percentage of them will show only once, trigger rejects the set if this percentage is higher than this value}

\item{ratio}{Rejects collumns that the ratio of unique values to total values is higher than this value, primary keys have ratio equal to 1}
}
\description{
\code{BestVectorTable} Asks for a data.table and some parameters and returns the best combination of collums to predict the missing value
}
