% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{candidates}
\alias{candidates}
\title{\code{candidates} Asks for a dataframe and some parameters and returns how close the collums chosen can predict the goal collum
Should be used mostly with generate_candidates or preferably best_vector in case you only want the best combination possible for prediction}
\usage{
candidates(df, goal, vec, repetitions, trigger = 1)
}
\arguments{
\item{df}{A dataframe with the missing values you wish to fill}

\item{goal}{The collum with the missing values you wish to fill}

\item{vec}{a vector of collums you wish to test if can be used to predict the values}

\item{repetitions}{Measure of error, the bigger the less likely you will get the right prediction}

\item{trigger}{If its 1 it wont use candidates that only show once, recomend to run always on 1  and on numeric values compare with 1 to 0}
}
\description{
\code{candidates} Asks for a dataframe and some parameters and returns how close the collums chosen can predict the goal collum
Should be used mostly with generate_candidates or preferably best_vector in case you only want the best combination possible for prediction
}
