% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{CompleteDatasetTable}
\alias{CompleteDatasetTable}
\title{\code{CompleteDatasetTable} Asks for a data.table, a vector of collumn indices and the goal collumn and  returns the data frame with the values filled}
\usage{
CompleteDatasetTable(df, rows, goal)
}
\arguments{
\item{df}{A data.table with the missing values you wish to fill}

\item{rows}{The collumns you wish to use to predict the missing values}

\item{goal}{The collum with the missing values you wish to fill}
}
\description{
\code{CompleteDatasetTable} Asks for a data.table, a vector of collumn indices and the goal collumn and  returns the data frame with the values filled
}
