% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{generate_candidates}
\alias{generate_candidates}
\title{\code{generate_candidates} Asks for a dataframe and some parameters and returns all possible combinations of collums for prediction that satisfy a given error in input
in a list the first element of the list are the combinations while the second is its measure of error,to get the best parameters call best_vector}
\usage{
generate_candidates(df, goal, maxi, repetitions, trigger = 1)
}
\arguments{
\item{df}{A dataframe with the missing values you wish to fill}

\item{goal}{The collum with the missing values you wish to fill}

\item{maxi}{What will be the length of possible combinations you will test example if 2 they will test up to all possible pairs of collums}

\item{repetitions}{Measure of error, the bigger the less likely you will get the right prediction}

\item{trigger}{If its 1 it wont use candidates that only show once, recomend to run always on 1  and on numeric values compare with 1 to 0}
}
\description{
\code{generate_candidates} Asks for a dataframe and some parameters and returns all possible combinations of collums for prediction that satisfy a given error in input
in a list the first element of the list are the combinations while the second is its measure of error,to get the best parameters call best_vector
}
