% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{GenerateCandidatesTable}
\alias{GenerateCandidatesTable}
\title{\code{GenerateCandidatesTable} Asks for a data.table and some parameters and returns all possible combinations of collums for prediction that satisfy a given error in input
in a list the first element of the list are the combinations while the second is its measure of error,to get the best parameters call BestVector}
\usage{
GenerateCandidatesTable(df, goal, maxi, repetitions, trigger = 1)
}
\arguments{
\item{df}{A data.table with the missing values you wish to fill}

\item{goal}{The collum with the missing values you wish to fill}

\item{maxi}{What will be the length of possible combinations you will test example if 2 they will test up to all possible pairs of collums}

\item{repetitions}{Measure of error, the bigger the less likely you will get the right prediction}

\item{trigger}{When you pair all possible combination of tuples a percentage of them will show only once, trigger rejects the set if this percentage is higher than this value}
}
\description{
\code{GenerateCandidatesTable} Asks for a data.table and some parameters and returns all possible combinations of collums for prediction that satisfy a given error in input
in a list the first element of the list are the combinations while the second is its measure of error,to get the best parameters call BestVector
}
