% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{MeanAccuracy}
\alias{MeanAccuracy}
\title{\code{MeanAccuracy} Asks for a dataframe, a vector of collumn indices and the goal collumn the expected value of accuracy of filling missing values if the dataset is representative}
\usage{
MeanAccuracy(df, colunas, goal)
}
\arguments{
\item{df}{A dataframe that you intend to fill missing values, warning this dataframe shall contain no missing values so the user must drop the lines it happens}

\item{colunas}{The collumns you wish to use to predict the missing values}

\item{goal}{The collum with the missing values you wish to fill}
}
\description{
\code{MeanAccuracy} Asks for a dataframe, a vector of collumn indices and the goal collumn the expected value of accuracy of filling missing values if the dataset is representative
}
