% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{MeanAccuracy_table}
\alias{MeanAccuracy_table}
\title{\code{MeanAccuracy_table} Asks for a data.table, a vector of collumn indices and the goal collumn the expected value of accuracy of filling missing values if the dataset is representative}
\usage{
MeanAccuracy_table(df, VECTORS, goal)
}
\arguments{
\item{df}{A data.table that you intend to fill missing values, warning this dataframe shall contain no missing values so the user must drop the lines it happens}

\item{VECTORS}{The collumns you wish to use to predict the missing values}

\item{goal}{The collum with the missing values you wish to fill}
}
\description{
\code{MeanAccuracy_table} Asks for a data.table, a vector of collumn indices and the goal collumn the expected value of accuracy of filling missing values if the dataset is representative
}
