% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{NA_VALUES}
\alias{NA_VALUES}
\title{\code{NA_VALUES} Asks for a dataframe and returns a table of how many missing values are in each collum}
\usage{
NA_VALUES(df)
}
\arguments{
\item{df}{A dataframe with the missing values you wish to fill}
}
\description{
\code{NA_VALUES} Asks for a dataframe and returns a table of how many missing values are in each collum
}
