% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{WorstAccuracy_table}
\alias{WorstAccuracy_table}
\title{\code{WorstAccuracy_table} Asks for a data.table, a vector of collumn indices and the goal collumn and returns the minimum possible value of accuracy of filling missing values}
\usage{
WorstAccuracy_table(df, VECTORS, goal)
}
\arguments{
\item{df}{A data.table that you intend to fill missing values, warning this dataframe shall contain no missing values so the user must drop the lines it happens}

\item{VECTORS}{The collumns you wish to use to predict the missing values}

\item{goal}{The collum with the missing values you wish to fill}
}
\description{
\code{WorstAccuracy_table} Asks for a data.table, a vector of collumn indices and the goal collumn and returns the minimum possible value of accuracy of filling missing values
}
