% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Encontrar_candidatos_dataset_v1.R
\name{Candidates}
\alias{Candidates}
\title{\code{Candidates} Asks for a dataframe and some parameters and returns how close the collums chosen can predict the goal collum
Should be used mostly with generate_candidates or preferably BestVector in case you only want the best combination possible for prediction}
\usage{
Candidates(df, goal, vec, repetitions, trigger = 1)
}
\arguments{
\item{df}{A dataframe with the missing values you wish to fill}

\item{goal}{The collum with the missing values you wish to fill}

\item{vec}{a vector of collums you wish to test if can be used to predict the values}

\item{repetitions}{Measure of error, the bigger the less likely you will get the right prediction}

\item{trigger}{When you pair all possible combination of tuples a percentage of them will show only once, trigger rejects the set if this percentage is higher than this value}
}
\description{
\code{Candidates} Asks for a dataframe and some parameters and returns how close the collums chosen can predict the goal collum
Should be used mostly with generate_candidates or preferably BestVector in case you only want the best combination possible for prediction
}
